requirements.txt
# LLM API 客户端
openai>=1.12.0
anthropic>=0.18.0
google-cloud-aiplatform>=1.43.0
transformers==4.38.2

# Web 界面
streamlit>=1.32.0

# 数据处理和可视化
pandas>=2.2.0
numpy==1.26.4
plotly>=5.19.0

# 工具库
python-dotenv>=1.0.0
aiohttp>=3.9.0
async-timeout>=4.0.0
pydantic>=2.6.0

# 测试
pytest==8.0.2
pytest-asyncio==0.23.5 